requirements.txt
pdfminer.six
python-docx == 0.8.11
openpyxl == 3.1.1
pandas == 1.5.3